+++ nacc_attribute_deriver/attributes/namespace/uds_namespace.py
@@ -2,9 +2,15 @@
 
 from datetime import date, datetime
 
-from nacc_attribute_deriver.attributes.namespace.namespace import FormNamespace
+from nacc_attribute_deriver.attributes.namespace.namespace import (
+    FormNamespace,
+    WorkingNamespace,
+)
 from nacc_attribute_deriver.symbol_table import SymbolTable
-from nacc_attribute_deriver.utils.errors import InvalidFieldError
+from nacc_attribute_deriver.utils.errors import (
+    AttributeDeriverError,
+    InvalidFieldError,
+)
 
 
 class UDSNamespace(FormNamespace):
@@ -37,6 +43,8 @@ class UDSNamespace(FormNamespace):
             raise InvalidFieldError(
                 f"Current file is not an UDS form: found {module}",
             )
+
+        self.__working = WorkingNamespace(table=table)
 
     def is_initial(self) -> bool:
         """Returns whether or not this is an initial packet."""
@@ -71,9 +79,24 @@ class UDSNamespace(FormNamespace):
     def generate_uds_dob(self) -> date:
         """Creates UDS DOB, which is used to calculate ages.
 
-        No birth day provided, so default to 1.
+        Grab from subject.info.working.cross-sectional.birthmo/birthyr
+        which is assumed to be the latest DOB. This is to account
+        for the fact that sometimes these values change over the
+        visits. Because many things rely on DOB, we pre-compute
+        this to get the latest. But if not found, default to
+        the file's metadata.
         """
-        birthmo = self.get_required("birthmo", int)
-        birthyr = self.get_required("birthyr", int)
+        birthmo = self.__working.get_cross_sectional_value(
+            "birthmo", int, default=self.get_required("birthmo", int)
+        )
+        birthyr = self.__working.get_cross_sectional_value(
+            "birthyr", int, default=self.get_required("birthyr", int)
+        )
 
+        if not birthmo or not birthyr:
+            raise AttributeDeriverError(
+                "Cannot determine UDS DOB: missing BIRTHMO and/or BIRTHYR"
+            )
+
+        # No birth day provided, so default to 1.
         return datetime(birthyr, birthmo, 1).date()